requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
backports.csv==1.0.7
certifi==2020.12.5
chardet==3.0.4
discord.py==1.6.0
dnspython==1.16.0
idna==2.10
multidict==5.1.0
praw==7.1.1
prawcore==1.5.0
protobuf==3.14.0
pymongo==3.11.2
python-dotenv==0.15.0
requests==2.25.1
schedule==0.6.0
six==1.15.0
tablib==3.0.0
typing-extensions==3.7.4.3
update-checker==0.18.0
urllib3==1.26.2
websocket-client==0.57.0
yarl==1.6.3
